requirements.txt
aiogram~=2.19
pillow~=9.0